==> code/01-oncoprint.Rmd <==
---
title: "01 - Oncoprint"
author: "Selin Jessa [[selin.jessa@mail.mcgill.ca](mailto:selin.jessa@mail.mcgill.ca)] and Bhavyaa Chandarana [[bhavyaa.chandarana@mail.mcgill.ca](mailto:bhavyaa.chandarana@mail.mcgill.ca)]"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output:
  html_document:
    keep_md: yes
    code_folding: show
    theme: flatly
    css: ../include/style.css
    toc: yes
    toc_depth: 4
    number_sections: true
    df_print: paged
    includes:
      before_body: ../include/header.html
      after_body:  ../include/footer.html
---

<!-- FRONT MATTER, insert configuration info -->

```{r header, echo = FALSE, results = 'asis', warning = FALSE}

# Index of the document
# ...determines name of the subfolder of `outputs` and `figures`
doc_id <- "01"

suppressMessages(library(here))

# Knit child document with header
res <- knitr::knit_child(here("include", "header.Rmd"), envir = environment(), quiet = TRUE)
cat(res, sep = '\n')

```

<!-- END OF FRONT MATTER -->


# Overview

Here we produce an oncoprint as an overview of the dataset. This code
is adapted from Steven Hébert.

# Libraries

```{r libs, warning = FALSE, message = FALSE, echo = TRUE, cache = FALSE}

knitr::opts_chunk$set(cache = FALSE)

library(here)
library(data.table)
library(magrittr)
library(tidyr)
library(dplyr)
library(readr)
library(glue)
library(ggplot2)
library(ComplexHeatmap)

source(here("include/style.R"))
source(here("code/functions/scRNAseq.R"))

ggplot2::theme_set(theme_min())

```

# Load metadata

```{r read_meta, cache = FALSE}

meta <- read_tsv(here("output/00/oncoprint.tsv"))

data <- as.data.table(meta)

# Show counts per group
table(data$Group)

```

# Set heatmap annotation

```{r, cache = FALSE}

palette_location <- c("Cerebellum"      = "goldenrod3",
                     "Hemispheric"     = "navy",
                     "Pons"            = "gold1",
                     "Posterior Fossa" = "orange",
                     "Thalamus"        = "darkolivegreen4",
                     "Unknown"         = "gray80")

palette_bulkRNAseq   <- c("Y" = "black",
                          "N" = "gray80")

palette_scRNAseq     <- c("Y" = "black",
                          "N" = "gray80")

palette_scMultiome   <- c("Y" = "black",
                          "N" = "gray80")

palette_newSample    <- c("Y" = "black",
                          "N" = "gray80")

palette_FOXR2positive <- c("Y" = "black",
                          "N" = "gray80")

palette_sex <- c("M" = "black",
                 "F" = "gray80",
                 "Unknown" = "gray50")

# Set the order of the samples
samples_order <- data$ID

# Add additional info (row) on the bottom of the oncoprint
bot_ann <- HeatmapAnnotation(
    annotation_name_gp = gpar(fontsize = 10),
    Group = data$Group,
    Location = data$Location_broad,
    FOXR2_positive = data$FOXR2_positive,
    Sex = data$Sex,
    New_sample = data$New_sample,
    bulkRNAseq = data$RNAseq,
    scRNAseq = data$scRNAseq,
    `Age(year)` = anno_points(as.matrix(as.numeric(data$Age)), border = F, size = unit(1, "mm")),
    col = list(Group = palette_groups,
               Location = palette_location,
               FOXR2_positive = palette_FOXR2positive,
               Sex = palette_sex,
               New_sample = palette_newSample,
               bulkRNAseq = palette_bulkRNAseq,
               scRNAseq = palette_scRNAseq),
    na_col = "white",
    gp = gpar(col = "white"))

```

# Create oncoprint

```{r oncoprint, fig.width = 14, fig.height = 10}

# plot just the heatmap annotation by including an empty 0-row heatmap
# following: https://github.com/jokergoo/ComplexHeatmap/issues/301#issuecomment-487481520
Heatmap(matrix(nc = 125, nr = 0),
        bottom_annotation = bot_ann)


```

<!-- END MATTER, insert reproducibility info -->

```{r footer, echo = FALSE, results = 'asis', warning = FALSE, cache = FALSE}

# Knit child document with header
res <- knitr::knit_child(here("include", "footer.Rmd"), envir = environment(), quiet = TRUE)
cat(res, sep = '\n')

```

<!-- END OF END MATTER -->
